% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationDates.R
\name{calculationDates}
\alias{calculationDates}
\title{Calculates dates with defined frequency with from and to inclusive}
\usage{
calculationDates(from, to, freq)
}
\arguments{
\item{from}{the starting date of the calculation dates}

\item{to}{the ending date of the calculation dates}

\item{freq}{the required frequency of calculations}
}
\value{
the function (calculationDates) returns a vector of the sequence of dates with
the defined frequency. The results sequence of dates with specific frequency
may be used for time series analysis
}
\description{
The functions take start and end dates, periodic frequency. The result is a vector of dates. The reason of creating this functions is to get dates that are sequence of dates ending at each ending date specified frequency. We could use seq.Date but it has two limitations. First it does provide periodic sequence but those sequences are on dependent on start or end date. Second, it excludes either the ending dates of specified frequency or the starting dates (when used in reverse mode). So even if we use endpoints from lubridate we will miss out one whole period.
For example if to=as.Date("2020-06-30") with by="quarter" then the second quarter of 2020 will skip the sequence of date created by seq.Date. Since we need all dates and periods including ending period, our function gets all dates of calculations
}
\examples{
calculationDates("2019-12-31", "2020-06-30", "daily")
calculationDates("2019-12-31", "2020-06-30", "weekly")
calculationDates("2019-12-31", "2020-06-30", "monthly")
calculationDates("2019-12-31", "2020-06-30", "quarterly")
calculationDates("2019-12-31", "2020-06-30", "yearly")
}
\seealso{
\link[base]{seq.Date}

\link[lubridate]{endpoints}
}
\author{
"Mulazim-Ali KHOKHAR \href{mailto:mulazim.ali.khokhar@vub.be}{mulazim.ali.khokhar@vub.be}"
}
