% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummiesRefinitiv.R
\name{dummiesRefinitiv}
\alias{dummiesRefinitiv}
\title{Create dummies for selected Index/Portfolio}
\usage{
dummiesRefinitiv(
  data,
  var_ID,
  var_InOut,
  var_Date,
  from = NULL,
  to = NULL,
  freq = "monthly"
)
}
\arguments{
\item{data}{A data frame. The file must contains following variables with specified data-type:
\itemize{
\item ID: The variable records Reuters Identification Code for each security (as character)
\item InOut: The variable records '+' or '-' indicating inclusion and exclusion of security from a Index/portfolio. The data-type for in_out must be a factor.
\item var_Date: The variable records dates. When read with 'in_out' variables clarifies in and out date of any security in the Index/Portfolio. Please make sure that the date variable has 'Date' format.
}}

\item{var_ID}{This is the security ID as per the data-set}

\item{var_InOut}{This is the variable with c("+", "-") for c(inclusion, exclusion) in the portfolio respectively}

\item{var_Date}{This is the variable of dates and read with var_InOu to define date of inclusion or exclusion.}

\item{from}{This is a constant date. The resulting time series of dummies will start from this date (or end of period depending on frequency selected). If it is left empty the function will take the starting date of the date in the data (min(data\link{,"date"})).}

\item{to}{This is a constant date. The resulting time series of dummies will end on this date (or end of period depending on frequency selected). If it is left empty the function will take the maximum date of the date in the data (max(data\link{,"date"})).}

\item{freq}{This is a constant string variable. It specifies the month as frequency as default. But you may choose any of c("daily", "weekly", 'monthly',"quarterly" or "yearly") as frequencies. The resulting dummies xts file will be of the same frequency as defined here.}
}
\description{
Create times series of dummies for selected Index/Portfolio from Refinitiv data set
}
\examples{
dummiesRefinitiv(
data=data.frame(ID = c("A.x", "B.v", "C.x", "B.v","D.x"),
                date = as.Date(c("2019-12-31","2019-12-31","2019-12-31","2020-03-31","2020-03-31")),
                in_out = c("+", "+", "+", "-", "+")),
var_ID="ID",
var_InOut="in_out",
var_Date = "date" ,
from="2019-12-31",
to="2020-06-30",
freq="monthly")
}
\author{
"Mulazim-Ali KHOKHAR \href{mailto:mulazim.ali.khokhar@vub.be}{mulazim.ali.khokhar@vub.be}"
}
