% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_CRSP.R
\docType{data}
\name{example_CRSP}
\alias{example_CRSP}
\title{example_CRSP}
\format{
A data frame of 4 rows and 3 columns
\describe{
\item{ID}{Numeric values giving unique identifier for an imaginary security}
\item{date_in}{Date values describing date of inclusion of imaginary security in imaginary portfolio}
\item{date_out}{Date values describing date of exclusion of imaginary security in imaginary portfolio}
}
}
\source{
Randomly generated data
}
\usage{
example_CRSP
}
\description{
A dataset containing randomly generated data to apply dummiesCRSP function.
}
\keyword{datasets}
