% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_Refinitiv.R
\docType{data}
\name{example_Refinitiv}
\alias{example_Refinitiv}
\title{example_Refinitiv}
\format{
A data frame of 5 rows and 3 columns
\describe{
\item{ID}{String values giving unique identifier for an imaginary security}
\item{date}{Date values describing date of inclusion/exclusion of imaginary security in imaginary portfolio when read with in_out variable}
\item{in_out}{String Values c("+","-") describing date of inclusion/exclusion respectively of imaginary security in imaginary portfolio}
}
}
\source{
Randomly generated data
}
\usage{
example_Refinitiv
}
\description{
A dataset containing randomly generated data to apply dummiesRefinitiv function.
}
\keyword{datasets}
