% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioConstituents.R
\docType{package}
\name{PortfolioConstituents}
\alias{PortfolioConstituents}
\alias{PortfolioConstituents-package}
\title{PortfolioConstituents}
\description{
Create time series of dummies for the constituents of Index/Portfolio from either CRSP or Refinitiv data.
}
\author{
\strong{Maintainer}: Mulazim-Ali KHOKHAR \email{mulazim.ali.khokhar@vub.be} (\href{https://orcid.org/0000-0001-9845-8519}{ORCID})

}
